fix: type

+++ tests/unit_tests/test_tf_ops.py
@@ -88,39 +88,39 @@ def test_nce():
     loss = NoiseConstrastiveEstimation()
     loss_ref = loss(
         logit_true=np.array([10.0, 10.0]).astype("float32"),
-        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]),
-        freq_true=np.array([0.1, 0.1]),
-        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]),
+        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]).astype("float32"),
+        freq_true=np.array([0.1, 0.1]).astype("float32"),
+        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]).astype("float32"),
     )
 
     loss_more = loss(
         logit_true=np.array([10.0, 10.0]).astype("float32"),
-        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]),
-        freq_true=np.array([0.8, 0.8]),
-        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]),
+        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]).astype("float32"),
+        freq_true=np.array([0.8, 0.8]).astype("float32"),
+        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]).astype("float32"),
     )
     assert loss_more > loss_ref
 
     loss_more = loss(
         logit_true=np.array([10.0, 10.0]).astype("float32"),
-        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]),
-        freq_true=np.array([0.1, 0.1]),
-        freq_negative=np.array([[0.01, 0.1], [0.1, 0.01]]),
+        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]).astype("float32"),
+        freq_true=np.array([0.1, 0.1]).astype("float32"),
+        freq_negative=np.array([[0.01, 0.1], [0.1, 0.01]]).astype("float32"),
     )
     assert loss_more > loss_ref
 
     loss_less = loss(
         logit_true=np.array([12.0, 12.0]).astype("float32"),
-        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]),
-        freq_true=np.array([0.1, 0.1]),
-        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]),
+        logit_negative=np.array([[0.0, 0.0], [0.0, 0.0]]).astype("float32"),
+        freq_true=np.array([0.1, 0.1]).astype("float32"),
+        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]).astype("float32"),
     )
     assert loss_less < loss_ref
 
     loss_less = loss(
         logit_true=np.array([10.0, 10.0]).astype("float32"),
-        logit_negative=np.array([[-4.0, 0.0], [0.0, -4.0]]),
-        freq_true=np.array([0.1, 0.1]),
-        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]),
+        logit_negative=np.array([[-4.0, 0.0], [0.0, -4.0]]).astype("float32"),
+        freq_true=np.array([0.1, 0.1]).astype("float32"),
+        freq_negative=np.array([[0.1, 0.1], [0.1, 0.1]]).astype("float32"),
     )
     assert loss_less < loss_ref
